refactor(sentry): split the escalation contract out of project-core
scripts/sentry-triage-project-core.mjs sat at 998 lines against the 1000-line
hard cap its own suite pins, so the needs-human escalation rules could not grow
there. Move them — the non-decision placeholder set, the question normalizer and
isDecisionReadyQuestion — into scripts/sentry-triage-escalation-contract.mjs and
re-export from the verdict contract, so every caller keeps one import surface.

The new module imports only the text layer, which imports nothing, so it adds no
cycle and no third-party edge to the workflow runtime closure. It does join the
agent write wrapper's staged closure, so the staging step copies it and the
closure pin lists it.

Pure move: no behaviour change, and the sorted case-name set across the eight
covering suites is byte-identical before and after (561 cases).

Refs #1782.
