+++ .github/workflows/sentry-triage-agent.yml
@@ -427,6 +427,7 @@ jobs:
           for f in \
             sentry-triage-agent-comment.mjs \
             sentry-triage-project-core.mjs \
+            sentry-triage-escalation-contract.mjs \
             sentry-triage-text.mjs \
             sentry-mcp-broker.mjs; do
             install -m 0444 "scripts/${f}" "${tools_dir}/${f}"
